Revert "feat(core): add shutdown lifecycle hook to node executor (#27354)"
This reverts commit b5a93364c582df2f73621c9d69cfc5ff29494f64.
